Modified code to allow float entry for the job tab however issue exists where it keeps changing everything to 1.72 or 17.2.

+++ UpHole/OriginalCode/inc/DataManagers/RecordManager.h
@@ -75,7 +75,7 @@ typedef struct _NEWHOLE_INFO
     U_INT16 BoreholeNumber;
     U_INT16 StartingRecordNumber;
     U_INT16 EndingRecordNumber;
-    INT16 DefaultPipeLength;
+    float DefaultPipeLength;
     INT16 Declination;
     INT16 Toolface;
     INT16 DesiredAzimuth;
